Shifted encoder to faros_config and provided to_json method
