+++ config/faros_config/faros_config/__init__.py
@@ -1,3 +1,5 @@
+import ipaddress
+import json
 from pydantic import BaseModel
 from typing import Optional
 import yaml
@@ -6,6 +8,17 @@ from .network import NetworkConfig
 from .bastion import BastionConfig
 from .cluster import ClusterConfig
 from .proxy import ProxyConfig
+
+
+class PydanticEncoder(json.JSONEncoder):
+    def default(self, obj):
+        obj_has_dict = getattr(obj, "dict", False)
+        if obj_has_dict and callable(obj_has_dict):
+            return obj.dict(exclude_none=True)
+        elif isinstance(obj, ipaddress._IPAddressBase):
+            return str(obj)
+        else:
+            return json.JSONEncoder.default(self, obj)
 
 
 class FarosConfig(BaseModel):
@@ -20,3 +33,7 @@ class FarosConfig(BaseModel):
             config = yaml.safe_load(f)
 
         return cls.parse_obj(config)
+
+    def to_json(self) -> str:
+        return json.dumps(self, sort_keys=True, indent=4,
+                          separators=(',', ': '), cls=PydanticEncoder)
